Refactor: Main 컴포넌트 리팩토링 - WebcamScreen 컴포넌트 생성 - WebcamButton 컴포넌트 생성 - Question 컴포넌트 생성 - zustand 적용
