+++ package.json
@@ -16,7 +16,8 @@
     "react-scripts": "5.0.1",
     "react-webcam": "^7.0.1",
     "styled-components": "^5.3.5",
-    "web-vitals": "^2.1.4"
+    "web-vitals": "^2.1.4",
+    "zustand": "^4.0.0-rc.1"
   },
   "scripts": {
     "start": "react-scripts start",
